LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-button-hover-expected.txt



FAIL CSS Test: ::scroll-button changes on hover via :hover assert_equals: ::scroll-button changes on hover via :hover expected "rgb(0, 0, 255)" but got ""
